- rename `/src/builtin_aliases_deps.ts` to `/src/alias.ts`, and refactor import statements. - remove the no longer needed `/src/builtin_aliases.ts` submodule. - update `/src/mod.ts` so that it no longer exports the aliases from `alias.ts`, because that will prevent all alias exports from being treeshaken when bundling `mod.ts`.
