+++ src/lambda.ts
@@ -4,7 +4,7 @@
 */
 
 import { type BindableFunction, bindMethodToSelfByName } from "./binder.ts"
-import { date_now, dom_clearTimeout, dom_setTimeout, promise_resolve } from "./builtin_aliases_deps.ts"
+import { date_now, dom_clearTimeout, dom_setTimeout, promise_resolve } from "./alias.ts"
 import { HybridTree, HybridWeakMap, LimitedStack, type SimpleMap, StrongTree, TREE_VALUE_UNSET } from "./collections.ts"
 import { DEBUG } from "./deps.ts"
