+++ src/mod.ts
@@ -1,14 +1,13 @@
 /** auto export for all modules, except {@link "devdebug"}. <br>
  * `devdebug` pollutes the `globalThis` object whenever imported.
  * thus, anyone desiring this module should import it using `import {...} from "@oazmi/kitchensink/devdebug"`. <br>
- * also, {@link "builtin_aliases"} is not exported either to avoid namespace pollution in IDEs <br>
+ * also, {@link "alias"} is not exported either to avoid namespace pollution in IDEs <br>
 */
 
+// export * from "./alias.ts"
 export * from "./array2d.ts"
 export * from "./binder.ts"
 export * from "./browser.ts"
-export * from "./builtin_aliases_deps.ts"
-// export * from "./builtin_aliases.ts"
 export * from "./collections.ts"
 export * from "./cryptoman.ts"
 // export * from "./devdebug.ts"
